Refactor:공간 기록 버튼 컴포넌트 Organism으로 따로 생성
